Infix to prefix .c
A Prefix expression, is a mathematical notation in which operators precede their
operands, in contrast to the more common infix expression, in which operators are placed
between operands.

Prefix/postfix notation is especially popular for its innate ability to express the
intended order of operations without the need for parentheses and other precedence rules,
as are usually employed with infix notation. That is why the calculators we use may internally
be using prefix expressions to evaluate your input expression.

Consider the following order for operator precedence:
1. ^
2. /, *
3. +, -


Input Format

Each test case will have 1 line as input.
The line contains an integer n and a string(character array) “infix_exp” of size n with
a space separating them.
1 <= n <= 100
Each character in infix_exp will be either 0-9 or one of ^, /, +, -, ) or (
Each input will always be vaild. No need to check if input it invalid.

Output Format

The output should:
1. The equivalent prefix expression which is a string(character array)
2. The evaluation result after processing the prefix expression.

These both need to be outputted after separating with a space