Create Implement a queue using two stacks..c
You are supposed to use the two stacks provided to you via the boilerplate code.
These stacks are provided to you in the form of two integer arrays and their respective top variables enclosed in a structure.

The top of the stack is initialised to -1, indicating the stack is empty. 
In order to perform the PUSH operation, we just pre increment the TOP variable and store the intended element in that particular position in the stack. 
Similarly for performing the POP operation, we retrieve the element present in the TOP position and simply post decrement it. 


Note: 
You are required to use the stack1 for inserting an element onto the queue. 
The stack2 should only be used as an auxiliary stack and remain empty at all times. 
The validate operation provided as a part of the boilerplate checks for the validity of both the stacks. 
So MAKE SURE to use the expected implementation of the problem and not a workaround of the same. 

USING THE BOILERPLATE CODE PROVIDED BY US IS MANDATORY.
YOU WILL FAIL MOST OF THE TEST CASES IN CASE YOU FAIL TO DO SO.

Input Format

Each test case starts with a digit ‘n’, followed by n operations on the queue. 

The four operations of the queue are denoted with a letter each given below:
1) E (enqueue):
You need to enqueue the integer following E onto the rear end of the queue. 
This operation prints 1 upon successful enqueue.

2) D (dequeue): 
You need to return the integer at the front end of the queue and remove it from the queue.
This operation prints the dequeued element. It returns -1 if the queue is empty.

3) V (validate): 
This operation has already been implemented as a part of the boilerplate.
This operation pops ‘x’ elements from stack1 where ‘x’ is an integer followed by ‘V’. It returns the element present in the stack after ‘x’ pops. If the stack goes empty at any point of time, it just returns -1.
This operation also checks if the stack2 is empty.
The stack remains intact after the validate operation is performed.

4) P (peek): 
This operation returns the element present at the front end of the queue. It returns -1 if the queue is empty.

Output Format

You don’t need to print anything as a part of this lab exercise. 
Make sure to remove all the printfs and scanfs used for debugging.