Create Implement Bottom up heap construction.c
Create a min heap out of an array of N integers of a certain size S. You will be implementing a bottom up approach to this. 

A Min Heap is a Tree where the root node has the most minimum key in the tree. Every node is smaller than its children. 
The above definition holds true for all sub-trees in the tree. This is called the Min Heap property. It is important to note that heaps are not always sorted, the key condition that they follow is that the largest or smallest element is placed on the root node (top) depending if it is a Max or Min Heap. 

The array maintains a 0-based index.

Input Format

Each test case starts with an input of 2 lines. 
First line of input consists of an integer N, representing the size of the array, which is the number of elements that it holds. 
Second line consists of an array of space separated integers out of which you'll be constructing a min heap. 

Output Format

The output will be an array of space separated integers which is actually the level order traversal of the min heap printed using the print_heap method.

Testcase 1
Input

4
3 2 4 1

Output

1 2 4 3